prac3

+++ function-1-1.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 double arrayMin(double* array, int size) {
-    int min = array[0];
+    double min = array[0];
     for (int i=0 ;i<size;i++) {
         if (array[i]<min) {
             min = array[i];
